requirements.txt
boto3==1.38.0
botocore==1.38.0
jmespath==1.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
s3transfer==0.12.0
six==1.17.0
urllib3==2.4.0
